refactor: 도착지 정보 입력 페이지 리팩토링 진행
템플릿 코드와 로직 함수를 분리했습니다
