+++ haul-fe/src/views/Request/ChoiceDst/ChoiceDst.jsx
@@ -1,3 +1,6 @@
+import { useState, useRef, useEffect, useContext } from "react";
+import { reservationStore } from "../../../store/reservationStore.jsx";
+import { useNavigate } from "react-router-dom";
 import MobileLayout from "../../../components/MobileLayout/MobileLayout.jsx";
 import Margin from "../../../components/Margin/Margin.jsx";
 import Header from "../../../components/Header/Header.jsx";
@@ -7,12 +10,8 @@ import SearchMap from "../../../components/Map/SearchMap/SearchMap.jsx";
 import Input from "../../../components/Input/Input.jsx";
 import BottomButton from "../../../components/Button/BottomButton.jsx";
 import NavigationBar from "../../../components/NavigationBar/NavigationBar.jsx";
-import { useState, useRef, useEffect, useContext } from "react";
-import { reservationStore } from "../../../store/reservationStore.jsx";
-import { useNavigate } from "react-router-dom";
-import { UrlMap, ErrorMessageMap } from "../../../data/GlobalVariable.js";
-import { isPhoneNumber } from "../../../utils/helper.js";
-import ToastMaker from "../../../components/Toast/ToastMaker.jsx";
+import { UrlMap } from "../../../data/GlobalVariable.js";
+import { CheckSubmitDisabledFun, SumbitStore } from "./index.jsx";
 
 const ChoiceDst = () => {
   const navigation = useNavigate();
@@ -47,35 +46,14 @@ const ChoiceDst = () => {
   }, []);
 
   useEffect(() => {
-    CheckSubmitDisabled(CheckSubmitDisabledFun());
+    CheckSubmitDisabled(
+      CheckSubmitDisabledFun({
+        mapInfo: mapInfo,
+        inDstDetailAddress: inDstDetailAddress,
+        inDstTel: inDstTel
+      })
+    );
   }, [mapInfo]);
-
-  //이 페이지에서 원하는 값이 다 있는지 체크
-  function CheckSubmitDisabledFun() {
-    const isMapInfoFilled =
-      mapInfo.coordinate.latitude !== "" &&
-      mapInfo.coordinate.longitude !== "" &&
-      mapInfo.detailAddress !== "";
-    const isDstDetailAddressFilled = inDstDetailAddress.current.trim() !== "";
-    const isDstTelFilled = inDstTel.current.trim() !== "";
-    return !(isMapInfoFilled && isDstDetailAddressFilled && isDstTelFilled);
-  }
-
-  function SumbitStore() {
-    if (!isPhoneNumber(inDstTel.current)) {
-      ToastMaker({ type: "error", children: ErrorMessageMap.InvalidTelformat });
-      return;
-    }
-    setDstInfo({
-      dstName: mapInfo.name,
-      dstAddress: mapInfo.detailAddress,
-      dstLatitude: Number(mapInfo.coordinate.latitude),
-      dstLongitude: Number(mapInfo.coordinate.longitude),
-      dstDetailAddress: inDstDetailAddress.current,
-      dstTel: inDstTel.current
-    });
-    navigation(UrlMap.choiceLoadInfoPageUrl);
-  }
 
   return (
     <MobileLayout>
@@ -111,7 +89,13 @@ const ChoiceDst = () => {
         placeholder="상세주소를 입력해주세요."
         onChange={({ target: { value } }) => {
           inDstDetailAddress.current = value;
-          CheckSubmitDisabled(CheckSubmitDisabledFun());
+          CheckSubmitDisabled(
+            CheckSubmitDisabledFun({
+              mapInfo: mapInfo,
+              inDstDetailAddress: inDstDetailAddress,
+              inDstTel: inDstTel
+            })
+          );
         }}
       />
       <Margin height="20px" />
@@ -124,7 +108,13 @@ const ChoiceDst = () => {
         type="tel"
         onChange={({ target: { value } }) => {
           inDstTel.current = value;
-          CheckSubmitDisabled(CheckSubmitDisabledFun());
+          CheckSubmitDisabled(
+            CheckSubmitDisabledFun({
+              mapInfo: mapInfo,
+              inDstDetailAddress: inDstDetailAddress,
+              inDstTel: inDstTel
+            })
+          );
         }}
       />
       <Margin height="30px" />
@@ -132,7 +122,13 @@ const ChoiceDst = () => {
         role="main"
         disabled={submitDisabled}
         onClick={() => {
-          SumbitStore();
+          SumbitStore({
+            inDstTel: inDstTel,
+            inDstDetailAddress: inDstDetailAddress,
+            mapInfo: mapInfo,
+            setDstInfo: setDstInfo,
+            navigation: navigation
+          });
         }}
       >
         선택완료
